+++ Easydict/App/Localizable.xcstrings
@@ -427,6 +427,7 @@
       }
     },
     "baidu_translate" : {
+      "comment" : "Baidu Translate",
       "localizations" : {
         "en" : {
           "stringUnit" : {
@@ -2343,19 +2344,19 @@
         }
       }
     },
-    "service.configuration.baidu.api_missing.tips" : {
+    "service.configuration.api_missing.tips %@" : {
       "extractionState" : "manual",
       "localizations" : {
         "en" : {
           "stringUnit" : {
             "state" : "translated",
-            "value" : "The Secret Key API is currently in use, but either the App ID or Secret Key is empty. Please go to Settings > Service > Baidu Translate, fill in the required configuration information, or switch API types."
+            "value" : "The Secret Key API is currently in use, but either the App ID or Secret Key is empty. Please go to Settings > Service > %@, fill in the required configuration information, or switch API types."
           }
         },
         "zh-Hans" : {
           "stringUnit" : {
             "state" : "translated",
-            "value" : "当前正使用 Secret Key API，但是 App ID 或者 Secret Key 为空，请前往设置-服务-百度翻译，填写相关配置信息，或切换 API 类型"
+            "value" : "当前正使用 Secret Key API，但是 App ID 或者 Secret Key 为空，请前往设置-服务-%@，填写相关配置信息，或切换 API 类型"
           }
         }
       }
